move db tasks to separate file
