+++ fab/exceptions.py
@@ -1,0 +1,2 @@
+class PreindexNotFinished(Exception):
+    """Thrown when a preindex of the database is not finished in a certain time period"""
